perf(web): defer Folio editor and Streamdown plugins to lazy chunks
The DocxEditor + Yjs + utif2 + prosemirror-tables graph (~1067 KB raw /
~333 KB gz) was being pulled into the entry chunk via a static import
chain from `host.tsx`, `review-panel.tsx`, `docx-browser-editor.tsx`,
and `file-viewer-with-ai.tsx`. Rollup hoisted the shared dependency
into the entry as the lowest common ancestor of four lazy surfaces,
so every page load shipped the editor.

Break the chain with `React.lazy()` boundaries:

- Split `MessageResponse` (Streamdown + cjk/math/mermaid plugins) into
  `message-response.tsx`, lazy-loaded from `message.tsx` behind a
  `Suspense fallback={null}`.
- Split `FileViewerWithAI` into `file-viewer-with-ai.impl.tsx`,
  lazy-loaded from `file-viewer-with-ai.tsx`. The fallback renders the
  underlying viewer (`children`) so the file stays visible while the
  AI overlay chunk loads.
- Split `ReviewPanel` into `review-panel.impl.tsx`, lazy-loaded from
  `review-panel.tsx`.
- Convert the static `DocxBrowserEditor` import in
  `$viewId.document.tsx` to `lazy()` with a `DocxLoadingShell`
  Suspense fallback.

Rollup now bundles the Folio graph into a single lazy chunk that only
loads when the document route, peek viewer, inspector panel, or PDF
viewer mounts. Total eager preload drops from roughly 640 KB gz to
506 KB gz on the homepage / login / settings paths.
